Update README.md
Backpropagation is one of the many methods to train a Neural Network. It works by considering the deviation between the actual output and the desired output and adjusting the parameters(weights and biases) of the neural network accordingly. Every time the output is labeled an error in a supervised training phase, the information is sent backwards. Each weight is updated based on how much they were responsible for the deviation. This way, the neural network learns to minimize the chance of occurrence of errors. These occurrences of errors are represented by a loss function.